requirements.txt
# general modules
django >3.2, <3.3
python-decouple
pillow
djangorestframework

# third party modules
markdown
django-filter
drf-yasg[validation]

# email third party modules
django-mail-templated

# deployment modules

# refactoring and tests
black
flakc8
pytest
pytest-django